+++ src/pages/Detail/Components/Star.jsx
@@ -1,4 +1,5 @@
 import React, { useState, useEffect } from 'react';
+import StarIcon from './StarIcon';
 import './Star.scss';
 
 const Star = ({ max = 5, color = '#fcc419' }) => {
@@ -14,7 +15,7 @@ const Star = ({ max = 5, color = '#fcc419' }) => {
   if (Object.keys(product).length === 0) return null;
   return (
     <div className="icon">
-      <span className="star_icon">
+      <span>
         {Array.from({ length: max }, (_, i) => (
           <StarIcon
             key={i}
@@ -29,39 +30,3 @@ const Star = ({ max = 5, color = '#fcc419' }) => {
 };
 
 export default Star;
-
-const StarIcon = ({ color, filled }) => {
-  return (
-    <span>
-      {filled ? (
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          width="40"
-          height="39"
-          viewBox="0 0 14 13"
-        >
-          <path
-            id="star"
-            d="M9,2l2.163,4.279L16,6.969,12.5,10.3l.826,4.7L9,12.779,4.674,15,5.5,10.3,2,6.969l4.837-.69Z"
-            transform="translate(-2 -2)"
-            fill={color}
-          />
-        </svg>
-      ) : (
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          width="40"
-          height="39"
-          viewBox="0 0 14 13"
-        >
-          <path
-            id="star"
-            d="M9,2l2.163,4.279L16,6.969,12.5,10.3l.826,4.7L9,12.779,4.674,15,5.5,10.3,2,6.969l4.837-.69Z"
-            transform="translate(-2 -2)"
-            fill={'#f4f4f4'}
-          />
-        </svg>
-      )}
-    </span>
-  );
-};
