+++ src/pages/Detail/Components/StarIcon.jsx
@@ -1,0 +1,40 @@
+import React from 'react';
+import Star from './Star';
+
+const StarIcon = ({ color, filled }) => {
+  return (
+    <span>
+      {filled ? (
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          width="40"
+          height="39"
+          viewBox="0 0 14 13"
+        >
+          <path
+            id="star"
+            d="M9,2l2.163,4.279L16,6.969,12.5,10.3l.826,4.7L9,12.779,4.674,15,5.5,10.3,2,6.969l4.837-.69Z"
+            transform="translate(-2 -2)"
+            fill={color}
+          />
+        </svg>
+      ) : (
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          width="40"
+          height="39"
+          viewBox="0 0 14 13"
+        >
+          <path
+            id="star"
+            d="M9,2l2.163,4.279L16,6.969,12.5,10.3l.826,4.7L9,12.779,4.674,15,5.5,10.3,2,6.969l4.837-.69Z"
+            transform="translate(-2 -2)"
+            fill={'#f4f4f4'}
+          />
+        </svg>
+      )}
+    </span>
+  );
+};
+
+export default StarIcon;
