+++ src/pages/Detail/Detail.jsx
@@ -6,8 +6,8 @@ import deliveryFee from './배송정보.jpg';
 import './Detail.scss';
 
 const Detail = () => {
-  //  const params = useParams();
-  //  const productId = params.id;
+  const params = useParams();
+  const productId = params.id;
   const [product, setProduct] = useState({});
   const [count, setCount] = useState(1);
 
@@ -20,11 +20,28 @@ const Detail = () => {
   };
 
   useEffect(() => {
-    fetch('http://10.58.52.77:8000/products/2')
+    fetch('http://10.58.52.77:8000/products/${productId}')
       .then(res => res.json())
       .then(result => setProduct(result.data));
-  }, []);
+  }, [{ productId }]);
+
   const totalPrice = product.price * count;
+
+  const addCart = () =>
+    fetch('http://10.58.52.93:8000/cart', {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json;charset=utf-8',
+        authorization:
+          'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6Imxpa2VMaW9uODI4MkB3ZWNvZGUuY29tIiwidXNlcklkIjoxLCJpYXQiOjE2OTgxNjg3NTV9.8tjgbmwn2u7LeYuTKTjr3ZhTA1p5l0Nja5kUEs3yki4',
+      },
+      body: JSON.stringify({
+        productId: product.id,
+        quantity: count,
+      }),
+    })
+      .then(response => response.json())
+      .then(data => console.log(data));
 
   if (Object.keys(product).length === 0) return null;
 
@@ -68,11 +85,14 @@ const Detail = () => {
               <img className="deliveryFee" src={deliveryFee} alt="배송정보" />
               <div>배송비 50,000원 이상 무료배송</div>
               <br />
-              <input className="total" value="총 상품금액" />
+              <br />
+              <div className="total">총 상품금액</div>
               <div className="total">
                 {totalPrice.toLocaleString('ko-KR')}원
               </div>
-              <button className="cart">장바구니</button>
+              <button className="cart" onClick={addCart}>
+                장바구니
+              </button>
             </div>
           </div>
           <div className="detailBox" />
