+++ src/pages/Detail/Detail.scss
@@ -25,7 +25,7 @@
 
     .proudctImage {
       width: 500px;
-      height: 500px;
+      height: 550px;
       display: block;
     }
 
@@ -80,6 +80,7 @@
       border-radius: 6px;
       font-size: 25px;
       font-weight: bold;
+      text-align: center;
       height: 48px;
       color: white;
       background-color: rgb(0, 150, 243);
@@ -108,10 +109,9 @@
   margin-top: 20px;
   border-bottom: 1px solid lightgray;
   width: 100%;
-}
-
-.starIcon {
-  display: flex;
-  gap: 10px;
-  align-items: center;
+  .starIcon {
+    display: flex;
+    gap: 10px;
+    align-items: center;
+  }
 }
